requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
requests==2.32.4
PyMuPDF==1.26.3
faiss-cpu==1.11.0.post1
google-generativeai==0.8.5
